Smell summary and move vector smelling to a standalone script
